clippy: the hub's core to zero warnings, without weakening a lint
68 diagnostics across 22 files in server/src (66 warnings and 2 deny-level
errors) go to zero. The AI/curation half and server/crates belong to other
hands and are untouched.

Two were real, and both were latent rather than active:

  subsonic.rs playlist_json read `if with_entries || true` - a condition that
  could never be false. The rows are genuinely always needed (duration and the
  cover art come off them too), so the branch is gone and a comment says why.

  main.rs handle_cli_flags looped over the arguments, but every arm of the
  match ends in process::exit, so it only ever examined the first. Now it says
  so with `if let Some(arg) = args.first()`.

Three doc comments had come unstuck from their functions and were silently
documenting the wrong item: `disk_space` (api.rs), `record_play` and
`save_audio_features` (db.rs) each get theirs back.

The rest are tidying with identical behaviour: rsplitn(2) -> rsplit_once in
auth.rs's two token verifiers, repeat().take() -> repeat_n (3 in db.rs, 1 in
fx.rs), map_or(true,..) -> is_none_or, filter().next_back() -> rfind,
iter().any() -> contains, De Morgan on jams.rs's clock-handover guard (truth
table checked, all four cases identical), ScannedTrack fixtures built as
struct literals instead of default-then-reassign, two #[cfg(test)] modules
moved to the end of their file (provably pure reorderings - every non-blank
line is byte-identical), and six wide row tuples in db.rs plus one in
albums.rs replaced by named, documented type aliases so the column order lives
in one place.

Ten #[allow] attributes were added, every one with a written reason:

  seven clippy::too_many_arguments in db.rs, on writers whose argument list IS
  the column list of the row they upsert;
  one dead_code on subsonic_wire.rs's SubsonicError, because the enum is the
  Subsonic spec's error table and its discriminants are the wire numbers - a
  partial table is one nobody can check against the spec;
  one dead_code at the top of vibes.rs, naming the exact eight parked items it
  covers and pointing at the header that parks them. The module is otherwise
  LIVE (build_charts_reply, new_music_ids and key_for_seed all have callers in
  dj.rs and chartlists.rs), verified against the tree rather than trusted from
  a note.

No lint is weakened globally; there is no crate-level allow.

cargo test --workspace: 366 passed, 0 failed, 4 ignored (370 test functions,
unchanged). Verified on x86_64-unknown-linux-gnu as well as the host, because
mirror.rs's statvfs cast is the one change that could have been target-shaped.

+++ server/src/mirror.rs
@@ -116,7 +116,7 @@ pub(crate) fn free_bytes(path: &std::path::Path) -> Option<u64> {
                 return None;
             }
             // f_bavail: blocks an unprivileged process may actually use.
-            Some(stat.f_bavail as u64 * stat.f_frsize as u64)
+            Some(stat.f_bavail as u64 * stat.f_frsize)
         }
     }
     #[cfg(not(unix))]
